refactor: split search and invitation list
