+++ src/components/modules/ListUsersResult/ListUsersResult.scss
@@ -1,7 +1,7 @@
 @use '../../../assets/styles/variables' as v;
 
 .list-result {
-  max-height: 280px;
+  max-height: 630px;
   overflow: auto;
   background-color: v.$white;
   border-radius: 5px;
