+++ src/components/modules/NavDashboard/NavDashboard.tsx
@@ -7,6 +7,7 @@ import {
   SettingOutlined,
   UsergroupAddOutlined,
   UserOutlined,
+  UserAddOutlined,
 } from '@ant-design/icons';
 import Avatar from '@common/Avatar/Avatar';
 import Button from '@common/Button/Button';
@@ -65,6 +66,12 @@ const NavDashboard: React.FC = () => {
         icon: SearchOutlined,
         tooltipTitle: t('search'),
         key: ENavbar.SEARCH,
+        path: 'dashboard',
+      },
+      {
+        icon: UserAddOutlined,
+        tooltipTitle: t('invitation'),
+        key: ENavbar.INVITATION,
         path: 'dashboard',
       },
       {
