+++ src/components/modules/SidebarDashboard/SidebarDashboard.scss
@@ -5,4 +5,5 @@
   height: 100vh;
   margin-right: 4px;
   background-color: #f5f7fb;
+  overflow: hidden;
 }
