+++ src/components/modules/SidebarDashboard/SidebarDashboard.tsx
@@ -1,12 +1,13 @@
-import * as React from 'react';
-import { selectNavBar } from 'src/store/navbarSlice';
-import { useAppSelector } from 'src/store/hooks';
-import SidebarProfile from '@modules/SidebarProfile/SidebarProfile';
+import InvitationList from '@modules/InvitationList/InvitationList';
+import SidebarChats from '@modules/SidebarChats/SidebarChats';
 import SidebarContacts from '@modules/SidebarContacts/SidebarContacts';
 import SidebarGroups from '@modules/SidebarGroups/SidebarGroups';
-import SidebarChats from '@modules/SidebarChats/SidebarChats';
+import SidebarProfile from '@modules/SidebarProfile/SidebarProfile';
 import SidebarSearchUsers from '@modules/SidebarSearchUsers/SidebarSearchUsers';
+import * as React from 'react';
 import ENavbar from 'src/interfaces/ENavbar';
+import { useAppSelector } from 'src/store/hooks';
+import { selectNavBar } from 'src/store/navbarSlice';
 
 import './SidebarDashboard.scss';
 
@@ -20,6 +21,7 @@ const SidebarDashboard: React.FC = () => {
       {navbarAction === ENavbar.GROUPS && <SidebarGroups />}
       {navbarAction === ENavbar.CONTACTS && <SidebarContacts />}
       {navbarAction === ENavbar.SEARCH && <SidebarSearchUsers />}
+      {navbarAction === ENavbar.INVITATION && <InvitationList />}
     </div>
   );
 };
