+++ src/interfaces/ENavbar.tsx
@@ -4,6 +4,7 @@ enum ENavbar {
   GROUPS = 'GROUPS',
   CONTACTS = 'CONTACTS',
   SEARCH = 'SEARCH',
+  INVITATION = 'INVITATION',
   SETTINGS = 'SETTINGS',
 }
